fix(ledger): reject when ledger wallet
