+++ src/app/pages/rpc-sign-bip322-message/rpc-sign-bip322-message.tsx
@@ -1,6 +1,7 @@
 import { truncateMiddle } from '@stacks/ui-utils';
 
 import { useRouteHeader } from '@app/common/hooks/use-route-header';
+import { useRejectIfLedgerWallet } from '@app/common/rpc-helpers';
 import { Disclaimer } from '@app/components/disclaimer';
 import { NoFeesWarningRow } from '@app/components/no-fees-warning-row';
 import { PopupHeader } from '@app/features/current-account/popup-header';
@@ -25,6 +26,7 @@ export function RpcSignBip322MessageRoute() {
 
 function RpcSignBip322Message() {
   useRouteHeader(<PopupHeader displayAddresssBalanceOf="all" />);
+  useRejectIfLedgerWallet('signMessage');
 
   const {
     origin,
